+++ pkg/initcmd/view.go
@@ -6,7 +6,9 @@ import (
 	_ "embed"
 	"fmt"
 	"os"
+	"path"
 	"path/filepath"
+	"strings"
 	"text/template"
 
 	"github.com/airplanedev/cli/pkg/logger"
@@ -20,49 +22,68 @@ import (
 
 type InitViewRequest struct {
 	Prompter prompts.Prompter
-	Name     string
+	DryRun   bool
+
+	Name string
 }
 
-func InitView(ctx context.Context, req InitViewRequest) error {
+func InitView(ctx context.Context, req InitViewRequest) ([]string, error) {
+	filesCreated := []string{}
 	if req.Name == "" {
-		return errors.New("missing new view name")
+		return nil, errors.New("missing new view name")
 	}
 
 	slug := utils.MakeSlug(req.Name)
 	viewDir := ""
 
-	entrypoint, err := createViewEntrypoint(slug, req.Name)
+	entrypoint, err := createViewEntrypoint(req.DryRun, slug, req.Name)
 	if err != nil {
-		return err
+		return nil, err
 	}
+	filesCreated = append(filesCreated, entrypoint)
+
 	cwd, err := os.Getwd()
 	if err != nil {
-		return errors.Wrap(err, "getting working directory")
+		return nil, errors.Wrap(err, "getting working directory")
 	}
 	packageJSONDir := cwd
 	if fsx.Exists(filepath.Join(cwd, viewDir, "package.json")) {
 		packageJSONDir = filepath.Join(cwd, viewDir)
 	}
-	deps := []string{"@airplane/views", "react", "react-dom"}
-	deps = append(deps, "airplane")
-	packageJSONDir, err = node.CreatePackageJSON(packageJSONDir, node.PackageJSONOptions{
-		Dependencies: node.NodeDependencies{
-			Dependencies:    deps,
-			DevDependencies: []string{"@types/react", "@types/react-dom", "typescript"},
-		},
-	}, req.Prompter)
-	if err != nil {
-		return err
-	}
-	if filepath.Ext(entrypoint) == ".tsx" {
-		// Create/update tsconfig in the same directory as the package.json file
-		if err := node.CreateViewTSConfig(packageJSONDir, req.Prompter); err != nil {
-			return err
+	if !req.DryRun {
+		deps := []string{"@airplane/views", "react", "react-dom"}
+		deps = append(deps, "airplane")
+		packageJSONDir, err = node.CreatePackageJSON(packageJSONDir, node.PackageJSONOptions{
+			Dependencies: node.NodeDependencies{
+				Dependencies:    deps,
+				DevDependencies: []string{"@types/react", "@types/react-dom", "typescript"},
+			},
+		}, req.Prompter)
+		if err != nil {
+			return nil, err
 		}
 	}
+	filesCreated = append(filesCreated, path.Join(packageJSONDir, "package.json"))
 
-	if err := utils.CreateDefaultGitignoreFile(".gitignore"); err != nil {
-		return err
+	if filepath.Ext(entrypoint) == ".tsx" {
+		if !req.DryRun {
+			// Create/update tsconfig in the same directory as the package.json file
+			if err := node.CreateViewTSConfig(packageJSONDir, req.Prompter); err != nil {
+				return nil, err
+			}
+		}
+		filesCreated = append(filesCreated, path.Join(packageJSONDir, "tsconfig.json"))
+	}
+
+	if !req.DryRun {
+		if err := utils.CreateDefaultGitignoreFile(".gitignore"); err != nil {
+			return nil, err
+		}
+	}
+	filesCreated = append(filesCreated, ".gitignore")
+
+	if req.DryRun {
+		logger.Log("Running with --dry-run. This command would have created or updated the following files:\n- %s", strings.Join(filesCreated, "\n- "))
 	}
 
 	suggestNextViewSteps(suggestNextViewStepsRequest{
@@ -70,7 +91,7 @@ func InitView(ctx context.Context, req InitViewRequest) error {
 		slug:    slug,
 	})
 
-	return nil
+	return filesCreated, nil
 }
 
 func generateViewEntrypointPath(slug string) string {
@@ -80,7 +101,7 @@ func generateViewEntrypointPath(slug string) string {
 //go:embed view_scaffolding/Default.airplane.tsx
 var defaultViewEntrypointInline []byte
 
-func createViewEntrypoint(slug string, name string) (string, error) {
+func createViewEntrypoint(dryRun bool, slug string, name string) (string, error) {
 	entrypointPath := generateViewEntrypointPath(slug)
 
 	var entrypointContents []byte
@@ -97,8 +118,10 @@ func createViewEntrypoint(slug string, name string) (string, error) {
 		return "", errors.Wrap(err, "executing inline entrypoint template")
 	}
 	entrypointContents = buf.Bytes()
-	if err := os.WriteFile(entrypointPath, entrypointContents, 0644); err != nil {
-		return "", errors.Wrap(err, "creating view entrypoint")
+	if !dryRun {
+		if err := os.WriteFile(entrypointPath, entrypointContents, 0644); err != nil {
+			return "", errors.Wrap(err, "creating view entrypoint")
+		}
 	}
 	logger.Step("Created view entrypoint at %s", entrypointPath)
 	return entrypointPath, nil
